requirements.txt
pyserial
pillow
packbits
